Refactor: move providers
